SystemUI: open settings panel if there are no notifications
If there are no notifications that belong to the current user, then automatically open the quick settings menu (like the two-finger open). It only ignores the ADB icon for now (which is pretty useless) but could be extended to ignore the usb mounting one or maybe another?

TODO: add Settings.System toggle to disable the option

Change-Id: I3b161ae4fc985bcc26b22f4be9ab045036508278
Signed-off-by: Roman Birg <romanbirg@gmail.com>