Email force close when pressing back key after selecting characters
Email and Browser will be forced closed after selecting a text
and then press the back key, tilting the screen or trying to
share or copy the text. This is because a previouse fix
is using a part of the code that Google has removed.
Fix: clearSelection will now use its own EventHub id,
CLEAR_SELECT_TEXT.

This part describes the previous fix.
Marking a text on the web page and then press copy works,
but trying to mark the same text again does not work.
The reason for this is that the selection never gets
cleared in webkit.
The fix, calling cleaSelection in the onDestroyActionMode.
Also added clearSelection when getting an
onConfigurationChange.

FIX=DMS01503793

Conflicts:
	core/java/android/webkit/WebViewClassic.java

Change-Id: I5e5673bac4efb3d28e9fc613da4bf636f8ab3c39