Avoid layout passes when updating data activity indicators.
ImageView can optimize out layout passes if old and new drawable have
equal sizes. Take advantage of that by displaying a transparent drawable
with the same size as the in/out drawables when there's no activity
instead of clearing out the drawable in that case.

Restoring file and setting the difference

Trying to exclude the java file

Change-Id: Iaaa556ffab3ed4be1da8016ed40d0a9c2efcde01