fb: use on virtual POWER_KEY own global menu handling
we use the POWER_KEY down event to call the global menu from TGActions
to allow to add shorcuts to the global menu.

Handle this undependent of non virtual key events and call the global menu
directly without vibration and without delay.

Just put it into the message que and done

Change-Id: I7d9af04605696d41bee949cfd99f6de4c4be6baa