fb: handle globalmenu in TGActions not over Powerbutton anymore
We used till now a hacky workaround to show up the Power menu which worked
but is indeed not the best way. Last tweek on it broke 3rd party root application
which use similar hackish workaround.

Whatever make global menu over wm service available and call it from
TGActions directly. This respects of course the usual security checks
if the device is provisioned etc

Change-Id: Ia12ce467cee0d60741079b3fa2c78449d707a09f