Frameworks: Add user protection and pie/navbar toggle to TGActions
Add drawables for it and enable it for navbar, pie, hwkeys
and GlobalMenu and whatever other shorcutfeature want to have it

Keep the user save and do not allow to disable all navigations to keep
the device usable. User will be notified

In Settings user get the choice to enable the other navigation method. Of course
only on navbar devices due that hw key devices always have a working navigation

PS
- extend it and make the actions on global menu as toggle action
  to show on/off and reflect it on the icon as well

Change-Id: I69227e6653f4fe780aa41a91e3d4896d9ccc0f0a