Fix ResourceNotFoundException when updating  applied theme
If the user has a theme applied and there is an update to that theme,
then the themed apps will crash. This is because the theme's APK path changes,
but the apps do not update their resources. Apps were not updating their resources
because updateConfiguration() determined that the config (which only has pkgName) hasn't changed.

The problem is fixed by first applying the default theme config and then re-applying the
updated theme.

Conflicts:
	packages/SystemUI/src/com/android/systemui/quicksettings/UserTile.java

Change-Id: I00e32fc1535b4d279fcb37c35f868501ecc1dd19