Framework: Make the dialog divider color themeable
The holo blue color might not be suitable for some themes. This commit
swaps out the background of all titleDivider elements to use another
resource. This has no difference from the perspective of the user of a
"stock" build but this is huge for theme users.

Thanks to Arz Bhatia for the suggestion.

Change-Id: Icb917220accac3ad38770067120a7b1be894961a

Conflicts:
	core/res/res/layout/dialog_custom_title_holo.xml
	core/res/res/layout/dialog_title_holo.xml
	core/res/res/layout/dialog_title_icons_holo.xml